[hail] rip apart RowStore.scala
InputBuffer trait and implementors -> InputBuffers.scala
OutputBuffer trait and implementors -> OutputBuffers.scala
BufferSpec and implementors -> BufferSpecs.scala
CodecSpec -> CodecSpec.scala (renamed from CodecSpec2.scala)
MemoryBuffer class -> MemoryBuffer.scala
