+++ hail/src/main/scala/is/hail/io/RowStore.scala
@@ -10,7 +10,6 @@ import is.hail.expr.ir.{EmitFunctionBuilder, EmitUtils, EstimableEmitter, Method
 import is.hail.expr.types.MatrixType
 import is.hail.expr.types.physical._
 import is.hail.expr.types.virtual._
-import is.hail.io.compress.LZ4Utils
 import is.hail.io.fs.FS
 import is.hail.io.index.IndexWriter
 import is.hail.nativecode._
@@ -24,177 +23,6 @@ import org.apache.spark.sql.Row
 import org.apache.spark.{ExposedMetrics, TaskContext}
 import org.json4s.jackson.JsonMethods
 import org.json4s.{Extraction, JValue}
-
-trait BufferSpec extends Spec {
-  def buildInputBuffer(in: InputStream): InputBuffer
-
-  def buildOutputBuffer(out: OutputStream): OutputBuffer
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBuffer]
-
-  def buildCodeOutputBuffer(in: Code[OutputStream]): Code[OutputBuffer]
-
-  def nativeOutputBufferType: String
-
-  def nativeInputBufferType(inputStreamType: String): String
-}
-
-final case class LEB128BufferSpec(child: BufferSpec) extends BufferSpec {
-  def buildInputBuffer(in: InputStream): InputBuffer = new LEB128InputBuffer(child.buildInputBuffer(in))
-
-  def buildOutputBuffer(out: OutputStream): OutputBuffer = new LEB128OutputBuffer(child.buildOutputBuffer(out))
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBuffer] =
-    Code.newInstance[LEB128InputBuffer, InputBuffer](child.buildCodeInputBuffer(in))
-
-  def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBuffer] =
-    Code.newInstance[LEB128OutputBuffer, OutputBuffer](child.buildCodeOutputBuffer(out))
-
-  def nativeOutputBufferType: String = s"LEB128OutputBuffer<${ child.nativeOutputBufferType }>"
-
-  def nativeInputBufferType(inputStreamType: String): String =
-    s"LEB128InputBuffer<${ child.nativeInputBufferType(inputStreamType) }, $inputStreamType>"
-}
-
-final case class BlockingBufferSpec(blockSize: Int, child: BlockBufferSpec) extends BufferSpec {
-  require(blockSize <= (1 << 16))
-
-  def buildInputBuffer(in: InputStream): InputBuffer = new BlockingInputBuffer(blockSize, child.buildInputBuffer(in))
-
-  def buildOutputBuffer(out: OutputStream): OutputBuffer = new BlockingOutputBuffer(blockSize, child.buildOutputBuffer(out))
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBuffer] =
-    Code.newInstance[BlockingInputBuffer, Int, InputBlockBuffer](blockSize, child.buildCodeInputBuffer(in))
-
-  def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBuffer] =
-    Code.newInstance[BlockingOutputBuffer, Int, OutputBlockBuffer](blockSize, child.buildCodeOutputBuffer(out))
-
-  def nativeOutputBufferType: String = s"BlockingOutputBuffer<$blockSize, ${ child.nativeOutputBufferType }>"
-
-  def nativeInputBufferType(inputStreamType: String): String =
-    s"BlockingInputBuffer<$blockSize, ${ child.nativeInputBufferType(inputStreamType) }, $inputStreamType>"
-}
-
-trait BlockBufferSpec extends Spec {
-  def buildInputBuffer(in: InputStream): InputBlockBuffer
-
-  def buildOutputBuffer(out: OutputStream): OutputBlockBuffer
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBlockBuffer]
-
-  def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBlockBuffer]
-
-  def nativeOutputBufferType: String
-
-  def nativeInputBufferType(inputStreamType: String): String
-}
-
-final case class LZ4BlockBufferSpec(blockSize: Int, child: BlockBufferSpec) extends BlockBufferSpec {
-  require(blockSize <= (1 << 16))
-
-  def buildInputBuffer(in: InputStream): InputBlockBuffer = new LZ4InputBlockBuffer(blockSize, child.buildInputBuffer(in))
-
-  def buildOutputBuffer(out: OutputStream): OutputBlockBuffer = new LZ4OutputBlockBuffer(blockSize, child.buildOutputBuffer(out))
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBlockBuffer] =
-    Code.newInstance[LZ4InputBlockBuffer, Int, InputBlockBuffer](blockSize, child.buildCodeInputBuffer(in))
-
-  def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBlockBuffer] =
-    Code.newInstance[LZ4OutputBlockBuffer, Int, OutputBlockBuffer](blockSize, child.buildCodeOutputBuffer(out))
-
-  def nativeOutputBufferType: String = s"LZ4OutputBlockBuffer<${ 4 + LZ4Utils.maxCompressedLength(blockSize) }, ${ child.nativeOutputBufferType }>"
-
-  def nativeInputBufferType(inputStreamType: String): String =
-    s"LZ4InputBlockBuffer<${ 4 + LZ4Utils.maxCompressedLength(blockSize) }, ${ child.nativeInputBufferType(inputStreamType) }, $inputStreamType>"
-}
-
-object StreamBlockBufferSpec {
-  def extract(jv: JValue): StreamBlockBufferSpec = new StreamBlockBufferSpec
-}
-
-final class StreamBlockBufferSpec extends BlockBufferSpec {
-  def buildInputBuffer(in: InputStream): InputBlockBuffer = new StreamBlockInputBuffer(in)
-
-  def buildOutputBuffer(out: OutputStream): OutputBlockBuffer = new StreamBlockOutputBuffer(out)
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBlockBuffer] =
-    Code.newInstance[StreamBlockInputBuffer, InputStream](in)
-
-  def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBlockBuffer] =
-    Code.newInstance[StreamBlockOutputBuffer, OutputStream](out)
-
-  def nativeOutputBufferType: String = s"StreamOutputBlockBuffer"
-
-  def nativeInputBufferType(inputStreamType: String): String = s"StreamInputBlockBuffer<$inputStreamType>"
-
-  override def equals(other: Any): Boolean = other.isInstanceOf[StreamBlockBufferSpec]
-}
-
-final class StreamBufferSpec extends BufferSpec {
-  override def buildInputBuffer(in: InputStream): InputBuffer = new StreamInputBuffer(in)
-
-  override def buildOutputBuffer(out: OutputStream): OutputBuffer = new StreamOutputBuffer(out)
-
-  def buildCodeInputBuffer(in: Code[InputStream]): Code[InputBuffer] =
-    Code.newInstance[StreamInputBuffer, InputStream](in)
-
-  def buildCodeOutputBuffer(out: Code[OutputStream]): Code[OutputBuffer] =
-    Code.newInstance[StreamOutputBuffer, OutputStream](out)
-
-  override def nativeOutputBufferType: String = s"StreamOutputBuffer"
-
-  override def nativeInputBufferType(inputStreamType: String): String = s"StreamInputBuffer<$inputStreamType>"
-
-  override def equals(other: Any): Boolean = other.isInstanceOf[StreamBufferSpec]
-}
-
-object CodecSpec {
-  val defaultBufferSpec: BufferSpec = LEB128BufferSpec(
-    BlockingBufferSpec(32 * 1024,
-      LZ4BlockBufferSpec(32 * 1024,
-        new StreamBlockBufferSpec)))
-
-  val default: CodecSpec = PackCodecSpec(defaultBufferSpec)
-
-  val defaultUncompressedBuffer: BufferSpec = BlockingBufferSpec(32 * 1024,
-    new StreamBlockBufferSpec)
-
-  val defaultUncompressed: CodecSpec = PackCodecSpec(defaultUncompressedBuffer)
-
-  val unblockedUncompressed: CodecSpec = PackCodecSpec(new StreamBufferSpec)
-
-  def fromShortString(s: String): CodecSpec = s match {
-    case "default" => CodecSpec.default
-    case "defaultUncompressed" => CodecSpec.defaultUncompressed
-    case "unblockedUncompressed" => CodecSpec.unblockedUncompressed
-  }
-
-  val baseBufferSpecs: Array[BufferSpec] = Array(
-    BlockingBufferSpec(64 * 1024,
-      new StreamBlockBufferSpec),
-    BlockingBufferSpec(32 * 1024,
-      LZ4BlockBufferSpec(32 * 1024,
-        new StreamBlockBufferSpec)),
-    new StreamBufferSpec)
-
-  val bufferSpecs: Array[BufferSpec] = baseBufferSpecs.flatMap { blockSpec =>
-    Array(blockSpec, LEB128BufferSpec(blockSpec))
-  }
-
-  val codecSpecs: Array[CodecSpec] = bufferSpecs.flatMap { bufferSpec =>
-    Array(PackCodecSpec(bufferSpec))
-  }
-
-  val supportedCodecSpecs: Array[CodecSpec] = bufferSpecs.flatMap { bufferSpec =>
-    Array(PackCodecSpec(bufferSpec))
-  }
-}
-
-
-
-trait CodecSpec extends Spec {
-  def makeCodecSpec2(pType: PType): CodecSpec2
-}
 
 case class PackCodecSpec(child: BufferSpec) extends CodecSpec {
   def makeCodecSpec2(pType: PType) = PackCodecSpec2(pType, child)
@@ -257,7 +85,6 @@ final case class PackCodecSpec2(eType: PType, child: BufferSpec) extends CodecSp
 }
 
 object ShowBuf {
-
   def apply(buf: Array[Byte], pos: Int, n: Int): Unit = {
     val sb = new StringBuilder()
     val len = if (n < 32) n else 32
@@ -282,798 +109,6 @@ object ShowBuf {
       j += 1
     }
     System.err.println(sb.toString())
-  }
-
-}
-
-trait OutputBlockBuffer extends Spec with Closeable {
-  def writeBlock(buf: Array[Byte], len: Int): Unit
-
-  def getPos(): Long
-}
-
-trait InputBlockBuffer extends Spec with Closeable {
-  def close(): Unit
-
-  def seek(offset: Long)
-
-  def readBlock(buf: Array[Byte]): Int
-}
-
-final class StreamBlockOutputBuffer(out: OutputStream) extends OutputBlockBuffer {
-  private val lenBuf = new Array[Byte](4)
-
-  def close() {
-    out.close()
-  }
-
-  def writeBlock(buf: Array[Byte], len: Int): Unit = {
-    Memory.storeInt(lenBuf, 0, len)
-    out.write(lenBuf, 0, 4)
-    out.write(buf, 0, len)
-  }
-
-  def getPos(): Long = out.asInstanceOf[ByteTrackingOutputStream].bytesWritten
-}
-
-final class StreamBlockInputBuffer(in: InputStream) extends InputBlockBuffer {
-  private val lenBuf = new Array[Byte](4)
-
-  def close() {
-    in.close()
-  }
-
-  // this takes a virtual offset and will seek the underlying stream to offset >> 16
-  def seek(offset: Long): Unit = in.asInstanceOf[ByteTrackingInputStream].seek(offset >> 16)
-
-  def readBlock(buf: Array[Byte]): Int = {
-    in.readFully(lenBuf, 0, 4)
-    val len = Memory.loadInt(lenBuf, 0)
-    assert(len >= 0)
-    assert(len <= buf.length)
-    in.readFully(buf, 0, len)
-    len
-  }
-}
-
-final class MemoryBuffer extends Serializable {
-  var mem: Array[Byte] = new Array[Byte](8)
-  var pos: Int = 0
-  var end: Int = 0
-
-  def capacity: Int = mem.length
-
-  def clear() {
-    pos = 0
-    end = 0
-  }
-
-  def clearPos() {
-    pos = 0
-  }
-
-  def grow(n: Int) {
-    mem = util.Arrays.copyOf(mem, math.max(capacity * 2, end + n))
-  }
-
-  def copyFrom(src: MemoryBuffer) {
-    mem = util.Arrays.copyOf(src.mem, src.capacity)
-    end = src.end
-    pos = src.pos
-  }
-
-  def writeByte(b: Byte) {
-    if (end + 1 > capacity)
-      grow(1)
-    Memory.storeByte(mem, end, b)
-    end += 1
-  }
-
-  def writeInt(i: Int) {
-    if (end + 4 > capacity)
-      grow(4)
-    Memory.storeInt(mem, end, i)
-    end += 4
-  }
-
-  def writeLong(i: Long) {
-    if (end + 8 > capacity)
-      grow(8)
-    Memory.storeLong(mem, end, i)
-    end += 8
-  }
-
-  def writeFloat(i: Float) {
-    if (end + 4 > capacity)
-      grow(4)
-    Memory.storeFloat(mem, end, i)
-    end += 4
-  }
-
-  def writeDouble(i: Double) {
-    if (end + 8 > capacity)
-      grow(8)
-    Memory.storeDouble(mem, end, i)
-    end += 8
-  }
-
-  def writeBytes(region: Region, off: Long, n: Int) {
-    if (end + n > capacity)
-      grow(n)
-    Memory.memcpy(mem, end, off, n)
-    end += n
-  }
-
-  def readByte(): Byte = {
-    assert(pos + 1 <= end)
-    val b = Memory.loadByte(mem, pos)
-    pos += 1
-    b
-  }
-
-  def readInt(): Int = {
-    assert(pos + 4 <= end)
-    val i = Memory.loadInt(mem, pos)
-    pos += 4
-    i
-  }
-
-  def readLong(): Long = {
-    assert(pos + 8 <= end)
-    val l = Memory.loadLong(mem, pos)
-    pos += 8
-    l
-  }
-
-  def readFloat(): Float = {
-    assert(pos + 4 <= end)
-    val f = Memory.loadFloat(mem, pos)
-    pos += 4
-    f
-  }
-
-  def readDouble(): Double = {
-    assert(pos + 8 <= end)
-    val d = Memory.loadDouble(mem, pos)
-    pos += 8
-    d
-  }
-
-  def readBytes(toRegion: Region, toOff: Long, n: Int) {
-    assert(pos + n <= end)
-    Memory.memcpy(toOff, mem, pos, n)
-    pos += n
-  }
-
-  def skipByte() {
-    assert(pos + 1 <= end)
-    pos += 1
-  }
-
-  def skipInt() {
-    assert(pos + 4 <= end)
-    pos += 4
-  }
-
-  def skipLong() {
-    assert(pos + 8 <= end)
-    pos += 8
-  }
-
-  def skipFloat() {
-    assert(pos + 4 <= end)
-    pos += 4
-  }
-
-  def skipDouble() {
-    assert(pos + 8 <= end)
-    pos += 8
-  }
-
-  def skipBytes(n: Int) {
-    assert(pos + n <= end)
-    pos += n
-  }
-}
-
-final class MemoryInputBuffer(mb: MemoryBuffer) extends InputBuffer {
-  def close() {}
-
-  def seek(offset: Long) = ???
-
-  def readByte(): Byte = mb.readByte()
-
-  def readInt(): Int = mb.readInt()
-
-  def readLong(): Long = mb.readLong()
-
-  def readFloat(): Float = mb.readFloat()
-
-  def readDouble(): Double = mb.readDouble()
-
-  def readBytes(toRegion: Region, toOff: Long, n: Int): Unit = mb.readBytes(toRegion, toOff, n)
-
-  def skipByte(): Unit = mb.skipByte()
-
-  def skipInt(): Unit = mb.skipInt()
-
-  def skipLong(): Unit = mb.skipLong()
-
-  def skipFloat(): Unit = mb.skipFloat()
-
-  def skipDouble(): Unit = mb.skipDouble()
-
-  def skipBytes(n: Int): Unit = mb.skipBytes(n)
-
-  def readDoubles(to: Array[Double], off: Int, n: Int): Unit = ???
-}
-
-final class MemoryOutputBuffer(mb: MemoryBuffer) extends OutputBuffer {
-  def flush() {}
-
-  def close() {}
-
-  def indexOffset(): Long = ???
-
-  def writeByte(b: Byte): Unit = mb.writeByte(b)
-
-  def writeInt(i: Int): Unit = mb.writeInt(i)
-
-  def writeLong(l: Long): Unit = mb.writeLong(l)
-
-  def writeFloat(f: Float): Unit = mb.writeFloat(f)
-
-  def writeDouble(d: Double): Unit = mb.writeDouble(d)
-
-  def writeBytes(region: Region, off: Long, n: Int): Unit = mb.writeBytes(region, off, n)
-
-  def writeDoubles(from: Array[Double], fromOff: Int, n: Int): Unit = ???
-}
-
-trait OutputBuffer extends Closeable {
-  def flush(): Unit
-
-  def close(): Unit
-
-  def indexOffset(): Long
-
-  def writeByte(b: Byte): Unit
-
-  def writeInt(i: Int): Unit
-
-  def writeLong(l: Long): Unit
-
-  def writeFloat(f: Float): Unit
-
-  def writeDouble(d: Double): Unit
-
-  def writeBytes(region: Region, off: Long, n: Int): Unit
-
-  def writeDoubles(from: Array[Double], fromOff: Int, n: Int): Unit
-
-  def writeDoubles(from: Array[Double]): Unit = writeDoubles(from, 0, from.length)
-
-  def writeBoolean(b: Boolean) {
-    writeByte(b.toByte)
-  }
-}
-
-final class StreamOutputBuffer(out: OutputStream) extends OutputBuffer {
-  private val buf = new Array[Byte](8)
-
-  override def flush(): Unit = out.flush()
-
-  override def close(): Unit = out.close()
-
-  def indexOffset(): Long = out.asInstanceOf[ByteTrackingOutputStream].bytesWritten
-
-  override def writeByte(b: Byte): Unit = out.write(Array(b))
-
-  override def writeInt(i: Int) {
-    Memory.storeInt(buf, 0, i)
-    out.write(buf, 0, 4)
-  }
-
-  def writeLong(l: Long) {
-    Memory.storeLong(buf, 0, l)
-    out.write(buf)
-  }
-
-  def writeFloat(f: Float) {
-    Memory.storeFloat(buf, 0, f)
-    out.write(buf, 0, 4)
-  }
-
-  def writeDouble(d: Double) {
-    Memory.storeDouble(buf, 0, d)
-    out.write(buf)
-  }
-
-  def writeBytes(region: Region, off: Long, n: Int): Unit = out.write(region.loadBytes(off, n))
-
-  def writeDoubles(from: Array[Double], fromOff: Int, n: Int) {
-    var i = 0
-    while (i < n) {
-      writeDouble(from(fromOff + i))
-      i += 1
-    }
-  }
-}
-
-final class LEB128OutputBuffer(out: OutputBuffer) extends OutputBuffer {
-  def flush(): Unit = out.flush()
-
-  def close() {
-    out.close()
-  }
-
-  def indexOffset(): Long = out.indexOffset()
-
-  def writeByte(b: Byte): Unit = out.writeByte(b)
-
-  def writeInt(i: Int): Unit = {
-    var j = i
-    do {
-      var b = j & 0x7f
-      j >>>= 7
-      if (j != 0)
-        b |= 0x80
-      out.writeByte(b.toByte)
-    } while (j != 0)
-  }
-
-  def writeLong(l: Long): Unit = {
-    var j = l
-    do {
-      var b = j & 0x7f
-      j >>>= 7
-      if (j != 0)
-        b |= 0x80
-      out.writeByte(b.toByte)
-    } while (j != 0)
-  }
-
-  def writeFloat(f: Float): Unit = out.writeFloat(f)
-
-  def writeDouble(d: Double): Unit = out.writeDouble(d)
-
-  def writeBytes(region: Region, off: Long, n: Int): Unit = out.writeBytes(region, off, n)
-
-  def writeDoubles(from: Array[Double], fromOff: Int, n: Int): Unit = out.writeDoubles(from, fromOff, n)
-}
-
-final class LZ4OutputBlockBuffer(blockSize: Int, out: OutputBlockBuffer) extends OutputBlockBuffer {
-  private val comp = new Array[Byte](4 + LZ4Utils.maxCompressedLength(blockSize))
-
-  def close() {
-    out.close()
-  }
-
-  def writeBlock(buf: Array[Byte], decompLen: Int): Unit = {
-    val compLen = LZ4Utils.compress(comp, 4, buf, decompLen)
-    Memory.storeInt(comp, 0, decompLen) // decompLen
-    out.writeBlock(comp, compLen + 4)
-  }
-
-  def getPos(): Long = out.getPos()
-}
-
-final class BlockingOutputBuffer(blockSize: Int, out: OutputBlockBuffer) extends OutputBuffer {
-  private val buf: Array[Byte] = new Array[Byte](blockSize)
-  private var off: Int = 0
-
-  def indexOffset(): Long = {
-    if (off == blockSize)
-      writeBlock()
-    (out.getPos() << 16) | off
-  }
-
-  private def writeBlock() {
-    out.writeBlock(buf, off)
-    off = 0
-  }
-
-  def flush() {
-    writeBlock()
-  }
-
-  def close() {
-    flush()
-    out.close()
-  }
-
-  def writeByte(b: Byte) {
-    if (off + 1 > buf.length)
-      writeBlock()
-    Memory.storeByte(buf, off, b)
-    off += 1
-  }
-
-  def writeInt(i: Int) {
-    if (off + 4 > buf.length)
-      writeBlock()
-    Memory.storeInt(buf, off, i)
-    off += 4
-  }
-
-  def writeLong(l: Long) {
-    if (off + 8 > buf.length)
-      writeBlock()
-    Memory.storeLong(buf, off, l)
-    off += 8
-  }
-
-  def writeFloat(f: Float) {
-    if (off + 4 > buf.length)
-      writeBlock()
-    Memory.storeFloat(buf, off, f)
-    off += 4
-  }
-
-  def writeDouble(d: Double) {
-    if (off + 8 > buf.length)
-      writeBlock()
-    Memory.storeDouble(buf, off, d)
-    off += 8
-  }
-
-  def writeBytes(fromRegion: Region, fromOff0: Long, n0: Int) {
-    assert(n0 >= 0)
-    var fromOff = fromOff0
-    var n = n0
-
-    while (off + n > buf.length) {
-      val p = buf.length - off
-      fromRegion.loadBytes(fromOff, buf, off, p)
-      off += p
-      fromOff += p
-      n -= p
-      assert(off == buf.length)
-      writeBlock()
-    }
-    fromRegion.loadBytes(fromOff, buf, off, n)
-    off += n
-  }
-
-  def writeDoubles(from: Array[Double], fromOff0: Int, n0: Int) {
-    assert(n0 >= 0)
-    assert(fromOff0 >= 0)
-    assert(fromOff0 <= from.length - n0)
-    var fromOff = fromOff0
-    var n = n0
-
-    while (off + (n << 3) > buf.length) {
-      val p = (buf.length - off) >>> 3
-      Memory.memcpy(buf, off, from, fromOff, p)
-      off += (p << 3)
-      fromOff += p
-      n -= p
-      writeBlock()
-    }
-    Memory.memcpy(buf, off, from, fromOff, n)
-    off += (n << 3)
-  }
-}
-
-trait InputBuffer extends Closeable {
-  def close(): Unit
-
-  def seek(offset: Long): Unit
-
-  def readByte(): Byte
-
-  def readInt(): Int
-
-  def readLong(): Long
-
-  def readFloat(): Float
-
-  def readDouble(): Double
-
-  def readBytes(toRegion: Region, toOff: Long, n: Int): Unit
-
-  def skipBoolean(): Unit = skipByte()
-
-  def skipByte(): Unit
-
-  def skipInt(): Unit
-
-  def skipLong(): Unit
-
-  def skipFloat(): Unit
-
-  def skipDouble(): Unit
-
-  def skipBytes(n: Int): Unit
-
-  def readDoubles(to: Array[Double], off: Int, n: Int): Unit
-
-  def readDoubles(to: Array[Double]): Unit = readDoubles(to, 0, to.length)
-
-  def readBoolean(): Boolean = readByte() != 0
-}
-
-final class StreamInputBuffer(in: InputStream) extends InputBuffer {
-  private val buff = new Array[Byte](8)
-
-  def close(): Unit = in.close()
-
-  def seek(offset: Long) = in.asInstanceOf[ByteTrackingInputStream].seek(offset)
-
-  def readByte(): Byte = {
-    in.read(buff, 0, 1)
-    Memory.loadByte(buff, 0)
-  }
-
-  def readInt(): Int = {
-    in.read(buff, 0, 4)
-    Memory.loadInt(buff, 0)
-  }
-
-  def readLong(): Long = {
-    in.read(buff)
-    Memory.loadLong(buff, 0)
-  }
-
-  def readFloat(): Float = {
-    in.read(buff, 0, 4)
-    Memory.loadFloat(buff, 0)
-  }
-
-  def readDouble(): Double = {
-    in.read(buff)
-    Memory.loadDouble(buff, 0)
-  }
-
-  def readBytes(toRegion: Region, toOff: Long, n: Int): Unit = {
-    toRegion.storeBytes(toOff, Array.tabulate(n)(_ => readByte()))
-  }
-
-  def skipByte(): Unit = in.skip(1)
-
-  def skipInt(): Unit = in.skip(4)
-
-  def skipLong(): Unit = in.skip(8)
-
-  def skipFloat(): Unit = in.skip(4)
-
-  def skipDouble(): Unit = in.skip(8)
-
-  def skipBytes(n: Int): Unit = in.skip(n)
-
-  def readDoubles(to: Array[Double], off: Int, n: Int): Unit = {
-    var i = 0
-    while (i < n) {
-      to(off + i) = readDouble()
-      i += 1
-    }
-  }
-}
-
-final class LEB128InputBuffer(in: InputBuffer) extends InputBuffer {
-  def close() {
-    in.close()
-  }
-
-  def seek(offset: Long): Unit = in.seek(offset)
-
-  def readByte(): Byte = {
-    in.readByte()
-  }
-
-  def readInt(): Int = {
-    var b: Byte = readByte()
-    var x: Int = b & 0x7f
-    var shift: Int = 7
-    while ((b & 0x80) != 0) {
-      b = readByte()
-      x |= ((b & 0x7f) << shift)
-      shift += 7
-    }
-    x
-  }
-
-  def readLong(): Long = {
-    var b: Byte = readByte()
-    var x: Long = b & 0x7fL
-    var shift: Int = 7
-    while ((b & 0x80) != 0) {
-      b = readByte()
-      x |= ((b & 0x7fL) << shift)
-      shift += 7
-    }
-    x
-  }
-
-  def readFloat(): Float = in.readFloat()
-
-  def readDouble(): Double = in.readDouble()
-
-  def readBytes(toRegion: Region, toOff: Long, n: Int): Unit = in.readBytes(toRegion, toOff, n)
-
-  def skipByte(): Unit = in.skipByte()
-
-  def skipInt() {
-    var b: Byte = readByte()
-    while ((b & 0x80) != 0)
-      b = readByte()
-  }
-
-  def skipLong() {
-    var b: Byte = readByte()
-    while ((b & 0x80) != 0)
-      b = readByte()
-  }
-
-  def skipFloat(): Unit = in.skipFloat()
-
-  def skipDouble(): Unit = in.skipDouble()
-
-  def skipBytes(n: Int): Unit = in.skipBytes(n)
-
-  def readDoubles(to: Array[Double], toOff: Int, n: Int): Unit = in.readDoubles(to, toOff, n)
-}
-
-final class LZ4InputBlockBuffer(blockSize: Int, in: InputBlockBuffer) extends InputBlockBuffer {
-  private val comp = new Array[Byte](4 + LZ4Utils.maxCompressedLength(blockSize))
-  private var decompBuf = new Array[Byte](blockSize)
-  private var lim = 0
-
-  def close() {
-    in.close()
-  }
-
-  def seek(offset: Long): Unit = in.seek(offset)
-
-  def readBlock(buf: Array[Byte]): Int = {
-    val blockLen = in.readBlock(comp)
-    val result = if (blockLen == -1) {
-      -1
-    } else {
-      val compLen = blockLen - 4
-      val decompLen = Memory.loadInt(comp, 0)
-      LZ4Utils.decompress(buf, 0, decompLen, comp, 4, compLen)
-      decompLen
-    }
-    lim = result
-    result
-  }
-}
-
-final class BlockingInputBuffer(blockSize: Int, in: InputBlockBuffer) extends InputBuffer {
-  private val buf = new Array[Byte](blockSize)
-  private var end: Int = 0
-  private var off: Int = 0
-
-  private def readBlock() {
-    assert(off == end)
-    end = in.readBlock(buf)
-    off = 0
-  }
-
-  private def ensure(n: Int) {
-    if (off == end)
-      readBlock()
-    assert(off + n <= end)
-  }
-
-  def close() {
-    in.close()
-  }
-
-  def seek(offset: Long): Unit = {
-    in.seek(offset)
-    off = end
-    readBlock()
-    off = (offset & 0xFFFF).asInstanceOf[Int]
-    assert(off <= end)
-  }
-
-  def readByte(): Byte = {
-    ensure(1)
-    val b = Memory.loadByte(buf, off)
-    off += 1
-    b
-  }
-
-  def readInt(): Int = {
-    ensure(4)
-    val i = Memory.loadInt(buf, off)
-    off += 4
-    i
-  }
-
-  def readLong(): Long = {
-    ensure(8)
-    val l = Memory.loadLong(buf, off)
-    off += 8
-    l
-  }
-
-  def readFloat(): Float = {
-    ensure(4)
-    val f = Memory.loadFloat(buf, off)
-    off += 4
-    f
-  }
-
-  def readDouble(): Double = {
-    ensure(8)
-    val d = Memory.loadDouble(buf, off)
-    off += 8
-    d
-  }
-
-  def readBytes(toRegion: Region, toOff0: Long, n0: Int) {
-    assert(n0 >= 0)
-    var toOff = toOff0
-    var n = n0
-
-    while (n > 0) {
-      if (end == off)
-        readBlock()
-      val p = math.min(end - off, n)
-      assert(p > 0)
-      toRegion.storeBytes(toOff, buf, off, p)
-      toOff += p
-      n -= p
-      off += p
-    }
-  }
-
-  def skipByte() {
-    ensure(1)
-    off += 1
-  }
-
-  def skipInt() {
-    ensure(4)
-    off += 4
-  }
-
-  def skipLong() {
-    ensure(8)
-    off += 8
-  }
-
-  def skipFloat() {
-    ensure(4)
-    off += 4
-  }
-
-  def skipDouble() {
-    ensure(8)
-    off += 8
-  }
-
-  def skipBytes(n0: Int) {
-    var n = n0
-    while (n > 0) {
-      if (end == off)
-        readBlock()
-      val p = math.min(end - off, n)
-      n -= p
-      off += p
-    }
-  }
-
-  def readDoubles(to: Array[Double], toOff0: Int, n0: Int) {
-    assert(toOff0 >= 0)
-    assert(n0 >= 0)
-    assert(toOff0 <= to.length - n0)
-    var toOff = toOff0
-    var n = n0
-
-    while (n > 0) {
-      if (end == off)
-        readBlock()
-      val p = math.min(end - off, n << 3) >>> 3
-      assert(p > 0)
-      Memory.memcpy(to, toOff, buf, off, p)
-      toOff += p
-      n -= p
-      off += (p << 3)
-    }
   }
 }
